+++ Packages/webxr-interactions/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Spectator Camera.
+- Mixed Reality Capture.
 
 ## [0.15.0] - 2022-06-04
 ### Changed